refactor: move usr repositories to its own folder
