+++ src/repositories/postgres/index.js
@@ -1,6 +1,6 @@
 // src/repositories/postgres/index.js
-export * from './create-user.js'
-export * from './get-user-by-id.js'
-export * from './update-user.js'
-export * from './get-user-by-email.js'
-export * from './delete-user.js'
+export * from './user/create-user.js'
+export * from './user/get-user-by-id.js'
+export * from './user/update-user.js'
+export * from './user/get-user-by-email.js'
+export * from './user/delete-user.js'
