+++ src/repositories/postgres/user/delete-user.js
@@ -1,11 +1,11 @@
-import { PostgresHelper } from '../../db/postgres/helper.js'
+import { PostgresHelper } from '../../../db/postgres/helper.js'
 
 export class PostgresDeleteUserRepository {
     async execute(userId) {
         const deleteUser = await PostgresHelper.query(
-        `DELETE FROM users 
+            `DELETE FROM users 
             WHERE id = $1 RETURNING *`,
-            [userId]
+            [userId],
         )
         return deleteUser[0]
     }
